logs support
given the server would run in the background and not be interacted with
until required, it'd be a good practice to store logs from every session
in a git ignored log directory.

run.sh is more for production time run where there is a compiled binary
in the project directory, unlike shell.nix which is "run script" for
during development. it updates all dependencies and runs a temp binary
from current source code status.